move function DoLogging to utils.go
